requirements_streamlit.txt
streamlit
requests
python-dotenv
google-cloud-aiplatform
langchain
langchain-google-vertexai
pymongo
redis
sentence-transformers
pydantic
typing-extensions